+++ src/Blazor.FileReader.ServerSide.Demo/Pages/_Host.cshtml
@@ -25,6 +25,6 @@
     <app>@(await Html.RenderComponentAsync<App>())</app>
 
     <script src="_framework/blazor.server.js"></script>
-    <script src="FileReaderComponent.js"></script>
+    <script src="_content/Blazor.FileReader/FileReaderComponent.js"></script>
 </body>
 </html>
