+++ src/Blazor.FileReader.ServerSide.Demo/Startup.cs
@@ -1,3 +1,11 @@
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.ResponseCompression;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
+using System.Linq;
+using System.Net.Mime;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -34,11 +42,11 @@ namespace Blazor.FileReader.ServerSide.Demo
             }
 
             app.UseHttpsRedirection();
-
+            
             app.UseStaticFiles();
-
+            
             app.UseRouting();
-
+            
             app.UseEndpoints(endpoints =>
             {
                 endpoints.MapBlazorHub();
