+++ src/Blazor.FileReader.Wasm.Demo/Blazor.FileReader.Wasm.Demo.csproj
@@ -22,7 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Blazor.FileReader.Wasm\Blazor.FileReader.Wasm.csproj" />
+    <ProjectReference Include="..\Blazor.FileReader\Blazor.FileReader.csproj" />
   </ItemGroup>
 
 </Project>
