+++ src/Blazor.FileReader.Wasm.Demo/Pages/Index.razor
@@ -7,9 +7,9 @@ Welcome to your new filestreaming app.<br />
 This demo reads a file without doing anything particular with it.<br />
 There are other demos in the menu.<br /><br />
 
-<input type="file" ref="inputElement" />
-<button onclick="@ReadFile" class="btn btn-primary">Read file</button>
-<button onclick="@ClearFile" class="btn btn-primary">Clear</button>
+<input type="file" @ref="@inputElement" />
+<button @onclick="@ReadFile" class="btn btn-primary">Read file</button>
+<button @onclick="@ClearFile" class="btn btn-primary">Clear</button>
 
 <br /><br />
 <textarea style="max-width: 100%;" cols="50" rows="20">@Output</textarea>
