+++ src/Blazor.FileReader.Wasm.Demo/wwwroot/index.html
@@ -150,6 +150,5 @@
       }(window.location))
     </script>
     <script src="_framework/blazor.webassembly.js"></script>
-    <script src="FileReaderComponent.js"></script>
 </body>
 </html>
