+++ src/Blazor.FileReader.Wasm/SetupExtension.cs
@@ -32,7 +32,7 @@ namespace Blazor.FileReader
 
             private MonoWebAssemblyJSRuntime MonoWebAssemblyJSRuntime => 
                 CurrentJSRuntime as MonoWebAssemblyJSRuntime;
-            public TRes InvokeUnmarshalled<T1, T2, TRes>(string identifier, T1 arg1, T2 arg2) => 
+            public TRes InvokeUnmarshalled<T1, T2, TRes>(string identifier, T1 arg1, T2 arg2) =>
                 MonoWebAssemblyJSRuntime.InvokeUnmarshalled<T1, T2, TRes>(identifier, arg1, arg2);
         }
     }
