+++ src/Blazor.FileReader/Blazor.FileReader.csproj
@@ -28,13 +28,13 @@
 
   <ItemGroup>
     <!-- .js/.css files will be referenced via <script>/<link> tags; other content files will just be included in the app's 'dist' directory without any tags referencing them -->
-    <EmbeddedResource Include="content\**\*.js" LogicalName="blazor:js:%(RecursiveDir)%(Filename)%(Extension)" />
-    <EmbeddedResource Include="content\**\*.css" LogicalName="blazor:css:%(RecursiveDir)%(Filename)%(Extension)" />
+    <EmbeddedResource Include="wwwroot\**\*.js" LogicalName="blazor:js:%(RecursiveDir)%(Filename)%(Extension)" />
+    
     <!--EmbeddedResource Include="content\**" Exclude="**\*.js;**\*.css" LogicalName="blazor:file:%(RecursiveDir)%(Filename)%(Extension)" />-->
   </ItemGroup>
 
   <ItemGroup>
-    <EmbeddedResource Remove="content\FileReaderComponent.ts" />
+    <EmbeddedResource Remove="script\FileReaderComponent.ts" />
   </ItemGroup>
 
   <ItemGroup>
@@ -48,8 +48,9 @@
 
   <ItemGroup>
     <Folder Include="Properties\" />
-    <Content Include="content\*.*">
+    <Content Include="script\*.*">
       <Pack>false</Pack>
     </Content>
+    <Folder Include="wwwroot\" />
   </ItemGroup>
 </Project>
