--- src/Blazor.FileReader/ReadFileAsyncErrorParams.cs
@@ -1,11 +1,0 @@
-﻿namespace Blazor.FileReader
-{
-    public partial class FileReaderJsInterop
-    {
-        public class ReadFileAsyncErrorParams
-        {
-            public string CallBackId { get; set; }
-            public string Exception { get; set; }
-        }
-    }
-}
