--- src/Blazor.FileReader/ReadFileMarshalledAsyncCallbackParams.cs
@@ -1,12 +1,0 @@
-﻿namespace Blazor.FileReader
-{
-    public partial class FileReaderJsInterop
-    {
-        public class ReadFileMarshalledAsyncCallbackParams
-        {
-            public string CallBackId { get; set; }
-
-            public string Data { get; set; }
-        }
-    }
-}
