+++ src/Blazor.FileReader/tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "noImplicitAny": false,
+    "noEmitOnError": true,
+    "outDir": "wwwroot",
+    "removeComments": false,
+    "sourceMap": true,
+    "target": "es5",
+    "lib": [
+      "dom", "es5", "es6", "es2015", "es2016", "es2017"
+    ]
+  },
+  "include": ["**/*.ts"],
+  "exclude": [
+    "node_modules",
+    "wwwroot"
+  ]
+}
